memoryhub-auth/pyproject.toml
[project]
name = "memoryhub-auth"
version = "0.1.0"
description = "OAuth 2.1 Authorization Server for MemoryHub"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "PyJWT>=2.9",
    "cryptography>=44.0",
    "bcrypt>=4.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov",
    "httpx>=0.27",
    "ruff",
    "aiosqlite",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
